requirements.txt
asgiref==3.8.1
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dj-database-url==2.3.0
Django==4.2.16
django-allauth==0.61.1
dj-rest-auth==5.0.1
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.8.2
flake8==6.1.0
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.5.1
stripe==11.2.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
